Fix types in test.

+++ dynamic_programming/test/test_sum_subarray_minimums.cpp
@@ -8,14 +8,14 @@
 TEST(ComputeSumSubArrayMinimums, Test1)
 {
 	std::vector<int> input{3, 1, 2, 4};
-	int result{17};
+	long result{17};
 	EXPECT_EQ(result, sum_of_subarray_minimums(input));
 }
 
 TEST(ComputeSumSubArrayMinimums, Test2)
 {
 	std::vector<int> input{11, 81, 94, 43, 3};
-	int result{444};
+	long result{444};
 	EXPECT_EQ(result, sum_of_subarray_minimums(input));
 }
 
@@ -23,6 +23,6 @@ TEST(ComputeSumSubArrayMinimums, Test2)
 TEST(ComputeSumSubArrayMinimums, Test3)
 {
 	std::vector<int> input{11, 2190, 2, 12, 190, 9, 23819, 3, 31, 12, 10190, 89, 679, 6730, 9890};
-	int result{62961};
+	long result{62961};
 	EXPECT_EQ(result, sum_of_subarray_minimums(input));
 }
